Check if provider changed on settingswindow exit Added custom signal handling in to handle this. Also moved *lots* of GpuPage functionality to imp()
Updated MainWindow UI:
- Renamed refresh button callback from "refresh_cards" to "refresh_cards_clicked"

Updated MainWindow class:
- Added a "update-all-views" signal with a new "signals()" function
- Added some better struct definition comments
- Split "refresh_cards()" function into two functions, one a pub function in MainWindow impl (so can be called by other functions) and "refresh_cards_clicked()" template callback
- Removed commented code from end of "refresh_cards()" function
- Added some better constructor comments
- Connected a closure to run on "update-all-views" signal, calls "refresh_cards()" function
- Updated SettingsWindow creation to pass self reference

Updated SettingsWindow class:
- Added a "ParentContainer" struct like in ModificationWindow class, functions pretty much the same
- Signal ("update-all-views") emission called from "close_request()" function instead
- Added some better constructor comments
- Added some better struct definition comments
- Updated constructor to handle new "ParentContainer" reference object

Updated GpuPage class:
- Moved "create_properties()" function to imp(), modified accordingly
- Moved "create_updater()" function to imp(), modified accordingly
- Moved "check_properties_for_view()" function to imp(), modified accordingly
- Added some better constructor comments
- Added some better struct definition comments
- Updated "load_views()" function to handle moved functions

Updated ModificationWindow class:
- Added some better struct definition comments
- Added some better constructor comments
- Removed some completed "TODO" comments
